GESTIONE [UPD] Allineamento gestione manutenzione con il nuovo modulo di caricamento delle pagine e puliti i sorgenti; correzione nel controllo permessi

+++ pages/gestione/manutenzione/missing.inc.php
@@ -1,4 +1,12 @@
 <?php
+
+/*HELP: */
+/*Controllo permessi utente*/
+if(!gdrcd_controllo_permessi($PARAMETERS['administration']['maintenance']['access_level'])) {
+    echo '<div class="error">'.gdrcd_filter('out', $MESSAGE['error']['not_allowed']).'</div>';
+    die();
+}
+
 if((is_numeric($_POST['mesi']) === true) && ($_POST['mesi'] >= 1) && ($_POST['mesi'] <= 12)) {
     /*Eseguo l'aggiornamento*/
     gdrcd_query("DELETE FROM clgpersonaggiooggetto WHERE nome IN (SELECT nome FROM personaggio WHERE DATE_SUB(NOW(), INTERVAL ".gdrcd_filter('num', $_POST['mesi'])." MONTH) > ora_entrata)");
@@ -17,7 +25,7 @@ if((is_numeric($_POST['mesi']) === true) && ($_POST['mesi'] >= 1) && ($_POST['me
     gdrcd_query("OPTIMIZE TABLE personaggio");
     ?>
     <!-- Conferma -->
-    <div class="warning">
+    <div class="success">
         <?php echo gdrcd_filter('out', $MESSAGE['warning']['modified']); ?>
     </div>
     <?php
@@ -27,10 +35,3 @@ if((is_numeric($_POST['mesi']) === true) && ($_POST['mesi'] >= 1) && ($_POST['me
     </div>
     <?php
 }
-?>
-<!-- Link di ritorno alla visualizzazione di base -->
-<div class="link_back">
-    <a href="main.php?page=gestione_manutenzione">
-        <?php echo gdrcd_filter('out', $MESSAGE['interface']['administration']['maintenance']['link']['back']); ?>
-    </a>
-</div>
